Move Code(Class)Wrapper to their own files
Prevents cycles in subsequent commit
