+++ amm/interp/src/main/scala/ammonite/interp/Interpreter.scala
@@ -3,7 +3,7 @@ package ammonite.interp
 import java.io.{File, OutputStream, PrintStream}
 import java.util.regex.Pattern
 
-import ammonite.interp.Preprocessor.CodeWrapper
+import ammonite.interp.CodeWrapper
 
 import scala.collection.mutable
 
@@ -39,8 +39,8 @@ class Interpreter(val printer: Printer,
                   verboseOutput: Boolean = true,
                   getFrame: () => Frame,
                   val createFrame: () => Frame,
-                  replCodeWrapper: Preprocessor.CodeWrapper,
-                  scriptCodeWrapper: Preprocessor.CodeWrapper,
+                  replCodeWrapper: CodeWrapper,
+                  scriptCodeWrapper: CodeWrapper,
                   alreadyLoadedDependencies: Seq[coursier.Dependency])
   extends ImportHook.InterpreterInterface{ interp =>
