+++ amm/src/main/scala/ammonite/Main.scala
@@ -3,8 +3,7 @@ package ammonite
 import java.io.{InputStream, OutputStream, PrintStream}
 import java.nio.file.NoSuchFileException
 
-import ammonite.interp.{Interpreter, Preprocessor}
-
+import ammonite.interp.{CodeClassWrapper, CodeWrapper, Interpreter, Preprocessor}
 import ammonite.runtime.{Frame, Storage}
 import ammonite.main._
 import ammonite.repl.Repl
@@ -66,8 +65,8 @@ case class Main(predefCode: String = "",
                 verboseOutput: Boolean = true,
                 remoteLogging: Boolean = true,
                 colors: Colors = Colors.Default,
-                replCodeWrapper: Preprocessor.CodeWrapper = Preprocessor.CodeWrapper,
-                scriptCodeWrapper: Preprocessor.CodeWrapper = Preprocessor.CodeWrapper,
+                replCodeWrapper: CodeWrapper = CodeWrapper,
+                scriptCodeWrapper: CodeWrapper = CodeWrapper,
                 alreadyLoadedDependencies: Seq[coursier.Dependency] =
                   Defaults.alreadyLoadedDependencies()){
 
@@ -403,9 +402,9 @@ class MainRunner(cliConfig: Cli.Config,
 
     val codeWrapper =
       if (cliConfig.classBased)
-        Preprocessor.CodeClassWrapper
+        CodeClassWrapper
       else
-        Preprocessor.CodeWrapper
+        CodeWrapper
 
     Main(
       cliConfig.predefCode,
